+++ Basil/include/VertexBuffer.h
@@ -5,6 +5,7 @@
 
 #include <cstdint>
 
+#include "BufferLayout.h"
 #include "Renderer.h"
 
 namespace Basil
@@ -15,6 +16,8 @@ namespace Basil
 			virtual ~VertexBuffer() {};
 			virtual void bind() const = 0;
 			virtual void unbind() const = 0;
+			virtual const BufferLayout& getLayout() const = 0;
+			virtual void setLayout(const BufferLayout& layout) = 0;
 			static VertexBuffer* create(std::vector<float>& vertices);
 	};
 }