+++ Basil/src/Application.cpp
@@ -9,6 +9,28 @@ namespace Basil
 	// Make Application a singleton
 	Application* Application::instance = nullptr;
 
+	// Convert shader data type to OpenGL type
+	static GLenum ShaderDataTypeToOpenGLBaseType(ShaderDataType type)
+	{
+		switch (type)
+		{
+			case ShaderDataType::Float:    return GL_FLOAT;
+			case ShaderDataType::Float2:   return GL_FLOAT;
+			case ShaderDataType::Float3:   return GL_FLOAT;
+			case ShaderDataType::Float4:   return GL_FLOAT;
+			case ShaderDataType::Mat3:     return GL_FLOAT;
+			case ShaderDataType::Mat4:     return GL_FLOAT;
+			case ShaderDataType::Int:      return GL_INT;
+			case ShaderDataType::Int2:     return GL_INT;
+			case ShaderDataType::Int3:     return GL_INT;
+			case ShaderDataType::Int4:     return GL_INT;
+			case ShaderDataType::Bool:     return GL_BOOL;
+		}
+
+		ASSERT(false, "Unknown ShaderDataType!");
+		return 0;
+	}
+
 	// Constructor
 	Application::Application()
 	{
@@ -32,9 +54,12 @@ namespace Basil
 		// Create vertex and index data
 		std::vector<float> vertices =
 		{
-			-0.5f, -0.5f, 0.0f,
-			 0.5f, -0.5f, 0.0f,
-			 0.0f,  0.5f, 0.0f
+			-0.5f, -0.5f, 0.0f,			// vertex 0 xyz
+			 1.0f,  0.0f, 0.0f, 1.0f,	// vertex 0 rgba
+			 0.5f, -0.5f, 0.0f,			// vertex 1 xyz
+			 0.0f,  1.0f, 0.0f, 1.0f,	// vertex 1 rgba
+			 0.0f,  0.5f, 0.0f,			// vertex 2 xyz
+			 0.0f,  0.0f, 1.0f, 1.0f	// vertex 2 rgba
 		};
 		
 		std::vector<unsigned int> indices =
@@ -46,23 +71,53 @@ namespace Basil
 		vbo.reset(VertexBuffer::create(vertices));
 		ibo.reset(IndexBuffer::create(indices));
 
+		// Set VBO's layout
+		{
+			BufferLayout layout =
+			{
+				{ "a_Position", ShaderDataType::Float3 },
+				{ "a_Color",	ShaderDataType::Float4 }
+			};
+
+			vbo->setLayout(layout);
+		}
+
 		// Create VAO and specify format of data
 		glGenVertexArrays(1, &vao);
 		glBindVertexArray(vao);
 		vbo->bind();
 		ibo->bind();
-		glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, 3 * sizeof(float), nullptr);
-		glEnableVertexAttribArray(0);
+
+		uint32_t index = 0;
+		const BufferLayout& layout = vbo->getLayout();
+		for (const BufferElement& element : layout)
+		{
+			glVertexAttribPointer(
+				index,
+				element.getComponentCount(),
+				ShaderDataTypeToOpenGLBaseType(element.type),
+				element.normalised ? GL_TRUE : GL_FALSE,
+				layout.getStride(),
+				(const void*)element.offset
+			);
+			glEnableVertexAttribArray(index);
+			index++;
+		}
 
 		// Create shaders
 		std::string vertexShaderSource = R"(
 			#version 330 core
 			
 			layout(location = 0) in vec3 a_Position;
+			layout(location = 1) in vec4 a_Color;
+			
 			out vec3 v_Position;
+			out vec4 v_Color;
+
 			void main()
 			{
 				v_Position = a_Position;
+				v_Color = a_Color;
 				gl_Position = vec4(a_Position, 1.0);	
 			}
 		)";
@@ -71,10 +126,13 @@ namespace Basil
 			#version 330 core
 			
 			layout(location = 0) out vec4 color;
+			
 			in vec3 v_Position;
+			in vec4 v_Color;
+			
 			void main()
 			{
-				color = vec4(v_Position * 0.5 + 0.5, 1.0);
+				color = v_Color;
 			}
 		)";